refactoring (ui)

+++ Frontend/src/app/customers/customers.ts
@@ -46,6 +46,9 @@ export class Customers implements OnInit {
   }
 
   handleDeleteCustomer(c: Customer) {
+    let conf = confirm("Are you sure you want to delete this customer ?");
+    if(!conf) return;
+
     this.customerService.deleteCustomer(c.id ).subscribe({
       next : (resp) => {
         this.customers = this.customers.pipe(
